Gui Figure Tweaks and Range Edit Box Functionality
The variables and text in the gui were redone so that there is text as a
title and text to be set to the variable value.
ex:
[Initial Water Level: ] [initialLevel]

Initial Water Level: will not be changed, but initialLevel is the tag of
the variable that will be changed.

Also, the range panel is working fully. Values are stored in
handles.minLevel and handles.maxLevel. The clear button sets these
values to 0 and clears the text boxes.